langchain/requirements.txt
langchain
langchain_community
langchain_openai
Flask
uvicorn
pydantic
python-dotenv